move all beacuse by miskte all are in page and in nextjs they treated as page so move all of them outside the page for the deployment is easy
