+++ src/pages/index.tsx
@@ -1,11 +1,11 @@
-import Footer from "@/pages/components/footer/Footer";
-import Review from "@/pages/components/review/Review";
-import Navbar from "@/pages/components/navbar/Navbar";
-import Gallary from "@/pages/Portfolio/Gallary";
-import HomeC from "@/pages/components/home/HomeC";
-import { sliderContent } from "@/pages/components/home/HomeContent";
-import Card from "@/pages/AboutUs/Card";
-import ServiceCard from "@/pages/Services/ServiceCard";
+import Footer from "@/components/footer/Footer";
+import Review from "@/components/review/Review";
+import Navbar from "@/components/navbar/Navbar";
+import Gallary from "@/components/Portfolio/Gallary";
+import HomeC from "@/components/home/HomeC";
+import { sliderContent } from "@/components/home/HomeContent";
+import Card from "@/components/AboutUs/Card";
+import ServiceCard from "@/components/Services/ServiceCard";
 export default function Home() {
   return (
     <div className="bg-slate-300">
